backend/requirements.txt
Flask
supabase
python-dotenv
phidata
openai
duckduckgo-search
newspaper3k
PyGithub
python-jose
python-multipart
uvicorn
gunicorn
asgiref
flask-cors
